+++ src/expandableCalendar/calendarProvider.js
@@ -1,9 +1,11 @@
 import _ from 'lodash';
-import PropTypes from 'prop-types';
 import React, {Component} from 'react';
 import {Animated, TouchableOpacity} from 'react-native';
+import PropTypes from 'prop-types';
 import XDate from 'xdate';
 
+import dateutils from '../dateutils';
+import {xdateToData} from '../interface';
 import styleConstructor from './style';
 import CalendarContext from './calendarContext';
 
@@ -13,15 +15,23 @@ const UPDATE_SOURCES = commons.UPDATE_SOURCES;
 const iconDown = require('../img/down.png');
 const iconUp = require('../img/up.png');
 
+/**
+ * @description: Calendar context provider component
+ * @example: https://github.com/wix/react-native-calendars/blob/master/example/src/screens/expandableCalendar.js
+ */
 class CalendarProvider extends Component {
+  static displayName = 'CalendarProvider';
+
   static propTypes = {
-    // Initial date in 'yyyy-MM-dd' format. Default = Date()
+    /** Initial date in 'yyyy-MM-dd' format. Default = Date() */
     date: PropTypes.any.isRequired,
-    // callback for date change event
+    /** callback for date change event */
     onDateChanged: PropTypes.func,
-    // whether to show the today button
+    /** callback for month change event */
+    onMonthChange: PropTypes.func,
+    /** whether to show the today button */
     showTodayButton: PropTypes.bool,
-    // The opacity for the disabled today button (0-1)
+    /** The opacity for the disabled today button (0-1) */
     disabledOpacity: PropTypes.number
   }
 
@@ -55,10 +65,17 @@ class CalendarProvider extends Component {
   };
 
   setDate = (date, updateSource) => {
+    const sameMonth = dateutils.sameMonth(XDate(date), XDate(this.state.date));
+
     this.setState({date, updateSource, buttonIcon: this.getButtonIcon(date)}, () => {
       this.animateTodayButton(date);
     });
+
     _.invoke(this.props, 'onDateChanged', date, updateSource);
+    
+    if (!sameMonth) {
+      _.invoke(this.props, 'onMonthChange', xdateToData(XDate(date)), updateSource);
+    }
   }
 
   setDisabled = (disabled) => {